% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{add_testthat}
\alias{add_testthat}
\title{Add testthat to WORCS Project}
\usage{
add_testthat(worcs_directory = ".", ...)
}
\arguments{
\item{worcs_directory}{PARAM_DESCRIPTION, Default: '.'}

\item{...}{Additional arguments passed to \link[usethis]{use_testthat}}
}
\value{
No return value. This function is called for its side effects.
}
\description{
Wraps \link[usethis]{use_testthat} to add integration testing to
a WORCS Project.
}
\examples{
if(requireNamespace("withr", quietly = TRUE) &
  requireNamespace("testthat", quietly = TRUE)){
withr::with_tempdir({
  usethis::create_project(path = ".", rstudio = FALSE, open = FALSE)
  add_testthat()
})
}
}
