% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installation.R
\name{check_worcs_installation}
\alias{check_worcs_installation}
\title{Check worcs dependencies}
\usage{
check_worcs_installation(what = "all")
}
\arguments{
\item{what}{Character vector indicating which dependencies to check. Default:
`"all"`. Options include: `c("dependencies", "git_user", "libgit2",
"git_init", "git_add", "git_commit", "SSH", "tinytex")`.}
}
\value{
Logical, indicating whether all checks passed or not.
}
\description{
This function checks that all worcs dependencies are correctly
installed, and suggests how to remedy any missing dependencies.
}
\examples{
check_worcs_installation("dependencies")
}
