% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{closed_data}
\alias{closed_data}
\title{Save closed data file}
\usage{
closed_data(data, codebook = TRUE)
}
\arguments{
\item{data}{A data.frame to save.}

\item{codebook}{Logical, indicating whether to render a codebook or not. If
set to \code{TRUE}, the default, a file called 'codebook.Rmd' is created and
rendered to 'codebook.md' for 'GitHub'.}
}
\value{
Returns \code{NULL} invisibly. This
function is called for its side effects.
}
\description{
This function saves a data.frame as a .csv file, stores a
checksum in 'data_checksum.txt', appends the .gitignore file to ignore
all '.csv' files, and writes a stem for 'data_cleaning.R'.
}
\examples{
the_test <- "closeddata"
old_wd <- getwd()
dir.create(file.path(tempdir(), the_test))
setwd(file.path(tempdir(), the_test))
closed_data(iris[1:10, ], codebook = FALSE)
setwd(old_wd)
unlink(file.path(tempdir(), the_test))
}
\seealso{
open_data
}
