% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{make_codebook}
\alias{make_codebook}
\title{Create codebook for a dataset}
\usage{
make_codebook(data, render_file = TRUE, file_name = "codebook.Rmd")
}
\arguments{
\item{data}{A data.frame for which to create a codebook.}

\item{render_file}{Logical. Whether or not to render the document.}

\item{file_name}{Character. File name to write the codebook \code{rmarkdown}
file to.}
}
\value{
\code{Logical}, indicating whether or not the operation was
succesful. This function is mostly called for its side effect of rendering an
\code{rmarkdown} codebook.
}
\description{
Creates a rudimentary 'Rmarkdown' codebook for a dataset, and
renders it to markdown for 'GitHub'. Users can customize the 'Rmarkdown'
document, for example, by adding a column with variable descriptions, or a
paragraph with details on the data collection procedure.
}
\examples{
library(rmarkdown)
library(knitr)
file_name <- tempfile("codebook", fileext = ".Rmd")
make_codebook(iris, file_name = file_name)
unlink(c(
  file_name,
  gsub("\\\\.Rmd", "\\\\.md", file_name),
  gsub("\\\\.Rmd", "\\\\.html", file_name),
  gsub("\\\\.Rmd", "_files", file_name)
), recursive = TRUE)
}
