% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{connect_github}
\alias{connect_github}
\title{Connect a project with a git repository to GitHub}
\usage{
connect_github(https)
}
\arguments{
\item{https}{https address of a GitHub repository, provided by GitHub when
creating a new Repository.}
}
\description{
This function adds the https address of a GitHub repository
as a remote to a local project, and pushes to local project to that remote
repository.
}
