% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_data.R
\name{data_unlabel}
\alias{data_unlabel}
\title{Drop value labels}
\usage{
data_unlabel(x, variables = names(x)[sapply(x, inherits, what = "factor")])
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{variables}{Column names of \code{x} to coerce to integer.}
}
\value{
A \code{data.frame}.
}
\description{
Coerces \code{factor} and \code{ordered} variables to class \code{integer}.
}
\examples{
\dontrun{
if(interactive()){
 df <- data.frame(x = factor(c("a", "b")))
 data_unlabel(df)
 }
}
}
