% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{descfun}
\alias{descfun}
\title{Describe a dataset}
\usage{
descfun(x)
}
\arguments{
\item{x}{An object for which a method exists.}
}
\value{
A data.frame
}
\description{
Provide descriptive statistics for a dataset.
}
\examples{
descfun(iris)
}
