% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_project}
\alias{export_project}
\title{Export Project}
\usage{
export_project(filename = NULL, open.data = TRUE)
}
\arguments{
\item{filename}{path to a target directory or path to a zip file that is to be created}

\item{open.data}{boolean. Export including open data?}
}
\value{
boolean. Success of the operation
}
\description{
Export Project
}
