% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{git_remote_connect}
\alias{git_remote_connect}
\title{Connect to Existing 'GitHub' Repository}
\usage{
git_remote_connect(repo, remote_repo)
}
\arguments{
\item{repo}{a path to an existing repository, or a git_repository object as returned by git_open, git_init or git_clone.}

\item{remote_repo}{Character, indicating the name of a repository on your
account.}
}
\value{
Invisibly returns a list with the following elements:
\itemize{
\item repo_url: Character, URL of the connected repository
\item repo_exists: Logical
\item prior_commits: Logical
}
}
\description{
Given that a 'GitHub' user is configured, with the appropriate
permissions, this function connects to an existing repository.
}
\examples{
\dontrun{
git_remote_connect()
}
}
\seealso{
\code{\link[gh]{gh_whoami}}
\code{\link[gert]{git_fetch}}, \code{\link[gert]{git_remote}}
}
