% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{github_action_reproduce}
\alias{github_action_reproduce}
\title{Set up GitHub Action to Reproduce WORCS Project}
\usage{
github_action_reproduce()
}
\value{
No return value. This function is called for its side effects.
}
\description{
Sets up a GitHub Action to perform continuous integration (CI)
for a WORCS project. CI automatically evaluates \code{reproduce()} and
\code{check_endpoints()} at each push or pull request.
}
\seealso{
\code{\link[usethis]{use_github_action}}
\code{\link[worcs]{add_endpoint}}
\code{\link[worcs]{check_endpoints}}
\code{\link[worcs]{github_action_check_endpoints}}
}
