% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{has_git_user}
\alias{has_git_user}
\title{Check whether global 'Git' credentials exist}
\usage{
has_git_user()
}
\value{
Logical, indicating whether 'Git' global configuration settings could
be retrieved, and contained the values
\code{user.name} and \code{user.email}.
}
\description{
Check whether the values \code{user.name} and \code{user.email}
exist exist in the 'Git' global configuration settings.
Uses \code{\link[gert]{git_config_global}}.
}
\examples{
has_git_user()
}
