% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report formatted number}
\usage{
report(x, digits = 2, equals = TRUE)
}
\arguments{
\item{x}{Numeric. Value to be reported}

\item{digits}{Integer. Number of digits to round to.}

\item{equals}{Logical. Whether to report an equals (or: smaller than) sign.}
}
\value{
An atomic character vector.
}
\description{
Report a number, rounded to a specific number of decimals (defaults to two),
using \code{\link{formatC}}. Intended for 'R Markdown' reports.
}
\examples{
if(interactive()){
report(.0234)
}
}
\author{
Caspar J. van Lissa
}
\keyword{internal}
