% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirnames_filenames.R
\name{worcs_root}
\alias{worcs_root}
\title{Return Absolute File Path of WORCS Project Directory}
\usage{
worcs_root(path = ".")
}
\arguments{
\item{path}{Start directory, Default: '.' (current directory).}
}
\value{
Normalized path of the \code{worcs} project root directory.
}
\description{
The search starts at \code{path}, and recursively proceeds up the
directory hierarchy until a \code{worcs} project directory is found.
}
\examples{
if(requireNamespace("withr", quietly = TRUE)){
withr::with_tempdir({
writeLines("", ".worcs")
worcs_root()
})
}
}
\seealso{
\code{\link[rprojroot]{find_root}}
}
