% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{cite_essential}
\alias{cite_essential}
\title{Essential citations Knit function for RStudio}
\usage{
cite_essential(inputFile, encoding)
}
\arguments{
\item{inputFile}{The file passed to \code{\link[rmarkdown]{render}}.}

\item{encoding}{Ignored. The encoding is always assumed to be UTF-8.}
}
\description{
This is a wrapper for \code{\link[rmarkdown]{render}}. First, this function
parses the citations in the document, removing citations
marked with double at sign, e.g.: \code{@@reference2020}.  Then, it renders
the file.
}
