% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{closed_data}
\alias{closed_data}
\title{Save closed data file}
\usage{
closed_data(data, codebook = "html")
}
\arguments{
\item{data}{A data.frame to save.}

\item{codebook}{A character string, indicating the filetype for the codebook.
Set to \code{NULL} to avoid creating a codebook. Defaults to \code{"html"},
other options documented in \code{\link[dataMaid]{makeDataReport}}.}
}
\description{
This function saves a data.frame as a .csv file, stores a
checksum in 'data_checksum.txt', appends the .gitignore file to ignore
all '.csv' files, and writes a stem for 'data_cleaning.R'.
}
\examples{
\dontshow{
the_test <- "closeddata"
old_wd <- getwd()
dir.create(file.path(tempdir(), the_test))
setwd(file.path(tempdir(), the_test))
closed_data(iris[1:5, ], codebook = NULL)
setwd(old_wd)
unlink(file.path(tempdir(), the_test))
}
\donttest{
closed_data(iris)
}
}
\seealso{
open_data
}
