% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{make_codebook}
\alias{make_codebook}
\title{Create codebook for a dataset}
\usage{
make_codebook(data, render_file = TRUE)
}
\arguments{
\item{data}{A data.frame for which to create a codebook.}

\item{render_file}{Logical. Whether or not to render the document.}
}
\description{
Creates a rudimentary Rmarkdown codebook for a dataset, and
renders it to markdown for GitHub. Users can customize the Rmarkdown
document, for example, by adding a column with variable descriptions, or a
paragraph with details on the data collection procedure.
}
\examples{
library(rmarkdown)
library(knitr)
old_wd <- getwd()
dir.create(file.path(tempdir(), "codebook"))
setwd(file.path(tempdir(), "citeall"))
make_codebook(iris)
setwd(old_wd)
unlink(file.path(tempdir(), "citeall"))
}
