% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{descriptives}
\alias{descriptives}
\title{Describe a dataset}
\usage{
descriptives(x)
}
\arguments{
\item{x}{An object for which a method exists.}
}
\value{
A \code{data.frame} with descriptive statistics for \code{x}.
}
\description{
Provide descriptive statistics for a dataset.
}
\examples{
descriptives(iris)
}
