% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_project}
\alias{export_project}
\title{Export project to .zip file}
\usage{
export_project(filename = NULL, open_data = TRUE)
}
\arguments{
\item{filename}{Character. Path to a \code{.zip} file that is to be created.
The default argument \code{NULL} creates a \code{.zip} file in the directory
one level above the 'worcs' project directory. By default, all files tracked
by 'Git' are included in the \code{.zip} file, excluding 'data.csv' if
\code{open_data = FALSE}.}

\item{open_data}{Logical. Whether or not to include the original data,
'data.csv', if this file exists. If \code{open_data = FALSE} and an open
data file does exist, then it is excluded from the \code{.zip} file. If it
does not yet exist, a synthetic data set is generated and added to the
\code{.zip} file.}
}
\value{
Logical, indicating the success of the operation. This function is
called for its side effect of creating a \code{.zip} file.
}
\description{
Export project to .zip file
}
\examples{
the_test <- "export"
old_wd <- getwd()
dir.create(file.path(tempdir(), the_test))
get_sig <- tryCatch(git_signature_default(), error = function(e){
  gert::git_config_global_set(name = "user.name", value = "yourname")
  gert::git_config_global_set(name = "user.email", value = "yourname@email.com")
})
suppressWarnings(worcs:::worcs_template(
                       file.path(tempdir(), the_test),
                       prereg_template = "none",
                       "add_license" = "none",
                       use_renv = FALSE,
                       remote_repo = "")
                       )
setwd(file.path(tempdir(), the_test))
result <- export_project(open_data = FALSE)
file.remove(normalizePath(file.path(tempdir(), paste0(the_test, ".zip"))))
setwd(old_wd)
unlink(file.path(tempdir(), the_test))
}
