% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{git_remote_create}
\alias{git_remote_create}
\title{Create a New 'GitHub' Repository}
\usage{
git_remote_create(name, private = TRUE)
}
\arguments{
\item{name}{Name of the repository to be created.}

\item{private}{Whether or not the repository should be private, defaults to
\code{FALSE}.}
}
\value{
Invisibly returns a logical value,
indicating whether the function was successful or not.
}
\description{
Given that a 'GitHub' user is configured, with the appropriate
permissions, this function creates a new repository on your account.
}
\examples{
git_remote_create()
}
