% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{load_data}
\alias{load_data}
\title{Load data file}
\usage{
load_data()
}
\value{
A \code{data.frame} of class \code{"worcs_data"}, containing the
original data, if available, or a synthetic version of the data, if the
original data are unavailable. The \code{data.frame} has an attribute, "type",
indicating whether the data are synthetic or original.
}
\description{
If the original data are available, this function loads the
original data. If only a synthetic dataset is available, this function loads
the synthetic data.
}
\examples{
the_test <- "loaddata"
old_wd <- getwd()
dir.create(file.path(tempdir(), the_test))
setwd(file.path(tempdir(), the_test))
open_data(iris[1:5, ], codebook = FALSE)
df <- load_data()
df
setwd(old_wd)
unlink(file.path(tempdir(), the_test))
}
