% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{load_data}
\alias{load_data}
\title{Load data file}
\usage{
load_data()
}
\value{
A data.frame.
}
\description{
If the original data are available, this function loads the
original data. If only a synthetic dataset is available, this function loads
the synthetic data.
}
\examples{
\dontshow{
the_test <- "loaddata"
old_wd <- getwd()
dir.create(file.path(tempdir(), the_test))
setwd(file.path(tempdir(), the_test))
open_data(iris[1:5, ], codebook = NULL)
load_data()
setwd(old_wd)
unlink(file.path(tempdir(), the_test))
}
\donttest{
open_data(iris[1:5, ], codebook = NULL)
load_data()
}
}
