% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{open_data}
\alias{open_data}
\title{Save open data file}
\usage{
open_data(data)
}
\arguments{
\item{data}{A data.frame to save.}
}
\description{
This function saves a data.frame as a .csv file, stores a
checksum in 'data_checksum.txt', appends the .gitignore file to exclude
'data.csv', and writes a stem for 'data_cleaning.R'.
}
\examples{
\dontrun{
  open_data(iris)
}
}
\seealso{
closed_data
}
