% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{open_data}
\alias{open_data}
\title{Save open data file}
\usage{
open_data(data, codebook = TRUE)
}
\arguments{
\item{data}{A data.frame to save.}

\item{codebook}{Logical, indicating whether to render a codebook or not. If
set to \code{TRUE}, the default, a file called 'codebook.Rmd' is created and
rendered to 'codebook.md' for GitHub.}
}
\description{
This function saves a data.frame as a .csv file, stores a
checksum in 'data_checksum.txt', appends the .gitignore file to exclude
'data.csv', and writes a stem for 'data_cleaning.R'.
}
\examples{
the_test <- "opendata"
old_wd <- getwd()
dir.create(file.path(tempdir(), the_test))
setwd(file.path(tempdir(), the_test))
open_data(iris[1:5, ], codebook = NULL)
setwd(old_wd)
unlink(file.path(tempdir(), the_test))
}
\seealso{
closed_data
}
