% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\name{reproduce}
\alias{reproduce}
\title{Reproduce WORCS Project}
\usage{
reproduce(worcs_directory = ".", verbose = TRUE, ...)
}
\arguments{
\item{worcs_directory}{Character, indicating the WORCS project directory
to which to save data. The default value "." points to the current directory.
Default: '.'}

\item{verbose}{Logical. Whether or not to print status messages to the
console. Default: \code{TRUE}}

\item{...}{Additional arguments.}
}
\value{
No return value. This function is called for its side effects.
}
\description{
Evaluate the recipe contained in a WORCS project to derive its
endpoints.
}
\examples{
# Create directory to run the example
old_wd <- getwd()
test_dir <- file.path(tempdir(), "reproduce")
dir.create(test_dir)
setwd(test_dir)
file.create(".worcs")
worcs:::add_recipe(recipe = 'writeLines("test", "test.txt")')
worcs:::reproduce()
# Cleaning example directory
setwd(old_wd)
unlink(test_dir, recursive = TRUE)
}
\seealso{
\code{\link[worcs]{add_endpoint}}
\code{\link[worcs]{snapshot_endpoints}}
\code{\link[worcs]{check_endpoints}}
}
