% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_ranger.R
\name{syn.ranger}
\alias{syn.ranger}
\title{Generating synthetic data sets using ranger}
\usage{
syn.ranger(y, x, xp, smoothing, proper = FALSE, ...)
}
\arguments{
\item{y}{Vector of the outcome variable.}

\item{x}{Matrix of x variables.}

\item{xp}{Matrix of x variables for prediction.}

\item{smoothing}{Logical. Whether of not to smooth predictions.}

\item{proper}{Whether or not to bootstrap the data, Default: FALSE}

\item{...}{Additional arguments passed to and from functions.}
}
\value{
List with predictions and fit model.
}
\description{
This function is for internal use only.
Generates synthetic version(s) of a data set using the fast
implementation of random forests in \code{\link[ranger]{ranger}}.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{internal}
